// src/interfaces/enums.ts
export enum THEME_MODE {
  LIGHT = 'light',
  DARK = 'dark',
}

export enum THEME_COLOR {
  GREEN = 'green',
  BLUE = 'blue',
  PURPLE = 'purple',
}

export enum INPUT_TYPES {
  NUMBER = 'number',
  EMAIL = 'email',
  PASSWORD = 'password',
  TEXT = 'text',
  CHECKBOX = 'checkbox',
  RADIO = 'radio',
  DATE = 'date',
  FILE = 'file',
  SELECT = 'select',
  EDITOR = 'editor',
  TAGS = 'tags',
}